refactor split cal_map.py in multiple .py files
